requirements.txt
python-telegram-bot
requests
click
pymongo
